SAE: Do not send PMKID to the driver if PMKSA caching is disabled
External auth status to the driver includes the PMKID derived as part of
SAE authentication, but this is not valid if PMKSA caching is disabled.
Drivers might not be expecting PMKID when it is not valid. Do not send
the PMKID to the driver in such cases.

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: 14d85a5af7fe74682033dbe50891c42e9ffa8c0d
Signed-off-by: Srinivas Dasari <dasaris@codeaurora.org>
Change-Id: I33f6ccd034bf2323924c74cfbd75dc002be84ffd
CRs-Fixed: 2421215