FILS: Verify RSNE match between Beacon/Probe Response and (Re)AssocResp
IEEE Std 802.11ai-2016 requires the FILS STA to do this check, but this
was missing from the initial implementation. The AP side behavior was
not described properly in 802.11ai due to a missing change in the
(Re)Association Response frame format tables which has resulted in some
deployed devices not including the RSNE.

For now, use an interoperability workaround to ignore the missing RSNE
and only check the payload of the element if it is present in the
protected frame. In other words, enforce this validation step only with
an AP that implements FILS authentication as described in REVmd while
allowing older implementations to skip this check (and the protection
against downgrade attacks). This workaround may be removed in the future
if it is determined that most deployed APs can be upgraded to add RSNE
into the (Re)Association Response frames.

Git-commit: 69afb6b395650ebc2fea405efb18d68ad855c1dd
Git-repo : git://w1.fi/srv/git/hostap.git
CRs-Fixed: 2465116
Change-Id: Id52ed1ec8df135aa229d591bd434c2aae9d58e9a
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>