hidl: route events to P2P device
The P2P device interface, when available, is used
for all P2P management operations, and all P2P events
are sent to it. The supplicant already routes all
relevant events to the P2P device but some events
are triggered on the network interface(group interface)
and not routed properly to the P2P device.
Fix the HIDL code to send these events on the P2P
device interface if found.

Change-Id: Icccf9738cb390b9976e74e00a1b9f8bfdd3937af
CRs-Fixed: 2504300