SAE: Allow PMKID to be added into Association Request frame following SAE
IEEE Std 802.11-2016 does not require this behavior from a SAE STA, but
it is not disallowed either, so it is useful to have an option to
identify the derived PMKSA in the immediately following Association
Request frames. This is disabled by default (i.e., no change to previous
behavior) and can be enabled with a global wpa_supplicant configuration
parameter sae_pmkid_in_assoc=1.

Change-Id: I35d033a4549bd67de42f02643fa958568956ccef
CRs-fixed: 2510423
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>