wpa_supplicant: Add support for EDMG channels
As part of a connection request, set EDMG parameters for the driver to
use in the association. The EDMG parameters are based on EDMG IE
received from the AP's Probe Response frames, the driver's EDMG
capability, and the EDMG configuration from the configuration.

As part of starting an AP, set EDMG parameters for the driver to use for
AP operation. The EDMG parameters are based on the driver's EDMG
capability and the EDMG configuration from the configuration.

This implementation is limited to CB2 (channel bonding of 2 channels)
and the bonded channels must be adjacent, i.e., the supported values for
edmg_channel are 9..13.

Signed-off-by: Alexei Avshalom Lazar <ailizaro@codeaurora.org>
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: a82aee1f4035bed9ceabae0108f7b0350b8482c9
Signed-off-by: Alexei Avshalom Lazar <ailizaro@codeaurora.org>
Change-Id: Ifcee6fd21849188b6d0a5c867a49386cd363af4e
CRs-Fixed: 2543749