OWE: Fix PTK derivation workaround for interoperability
The initial implementation of the PTK derivation workaround for
interoperability with older OWE implementations forced
WPA_KEY_MGMT_PSK_SHA256 to be used for all of PTK derivation. While that
is needed for selecting which hash algorithm to use, this was also
changing the length of the PTK components and by doing so, did not
actually address the backwards compatibility issue.

Fix this by forcing SHA256 as the hash algorithm in PTK derivation
without changing the PTK length calculation for OWE when
owe_ptk_workaround is enabled.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>

Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 0a76a0b96557a129fb11f17f84fdb6d7418dbe32
Change-Id: I4ee7b4fb2fca95f5c5a647c31925c6404323b81e
CRs-Fixed: 2748635