eap_proxy: Use IMSI encryption for anonymous@<realm> identity.
Framework configures anonymous identity as 'anonymous@<realm>' to
indicate supplicant that encrypted IMSI is to be used in-place of
original IMSI in EAP-Response/Identity when pseudonym or reauth-id
is not present.

This commit enables eap_proxy to configure 'use_encrypted_imsi' to
QMI, whenever anonymous identity is configured as 'anonymous@<realm>'.

Change-Id: Iaf1d05ce8a0911f410a92df08cdaacf2def5e3ca
CRs-Fixed: 2700598