WPS UPnP: Handle HTTP initiation failures for events more properly
While it is appropriate to try to retransmit the event to another
callback URL on a failure to initiate the HTTP client connection, there
is no point in trying the exact same operation multiple times in a row.
Replve the event_retry() calls with event_addr_failure() for these cases
to avoid busy loops trying to repeat the same failing operation.

These potential busy loops would go through eloop callbacks, so the
process is not completely stuck on handling them, but unnecessary CPU
would be used to process the continues retries that will keep failing
for the same reason.

Change-Id: I6d086a90ab9831ed4b050d37ce17fa941392cfb7
CRs-Fixed: 2702015
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 85aac526af8612c21b3117dadc8ef5944985b476
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>