P2P: Include channels 149 to 161 for operating classes 128 and 130
With the configuration defined in the global_op_class array, the
channels starting from 149 in the operating classes 128 and 130 were not
considered for P2P channel setup due to the non-continuous
incrementation of channel indexes. The other channels in these operating
classes were considered. Handle the channels from 149 to 161 in the
operating classes 128 and 130 for P2P channel setup by handling this
jump in the channel number incrementation.

CRs-fixed: 2729268
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 760d10cdea4d0c0e1a694e83b94b5dd0f71b7dc3
Change-Id: Ia8f6b1ae999b97f6882675f9fcc69f23141b22f5
Signed-off-by: Sreeramya Soratkal <ssramya@codeaurora.org>