wpa_supplicant: Consider bss entry pending for P2P joining as a known BSS
Consider the bss entry that is pending for P2P group join operation
also as known network along with the existing configured networks. This
prevents removal of bss entry that is still in the process of P2P
join operation from the BSS table when the number of entries exceed
bss_max_count.

CRs-fixed: 2795328
Signed-off-by: Sreeramya Soratkal <ssramya@codeaurora.org>
Change-Id: I6af2ee9eaf7e13ee9fa6661ea20625599cbd8df0