Cleanup on commit to consider pending P2P joining entry as Known BSS
This commit updates the delta changes to be in line with the
upstreamed commit. The upstreamed commit prevents removal of
the BSS entry that is still in the process of P2P join operation
from the BSS table when the number of entries exceed bss_max_count.

CRs-fixed: 2795328
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: c575904761b374e3738120580617c4a83a73e9f7
Change-Id: Ie5c259ddb33ace76b565e4e09613896d21035d8c