P2P: Include p2p_6ghz_disable in global configuration
Currently the configuration to disable 6GHz remains local to the
P2P interface. With this there is a possibility of 6GHz channels
being included in the channel list when the channel list needs to be
updated if state changes on one of the interfaces. Include the
configuration to disable 6GHz for P2P as global configuration to prevent
the inclusion of 6GHz channels in the channel list for P2P when channel
list needs to be updated during the state change in one of the interfaces.

CRs-Fixed: 2880565
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 2ffd3bb4b6dfd497fa1fb228b2cb4c6dafe37735
Signed-off-by: Sreeramya Soratkal <ssramya@codeaurora.org>
Change-Id: Ia8453f5875c5d0b7f09769963febb69af1bcb76e