FILS: Fix PMKID derivation for OKC
FILS authentication derives PMK differently from the EAP cases. The PMK
value does not bind in the MAC addresses of the STAs. As such, the same
PMKID is used with different BSSIDs. Fix both the hostapd and
wpa_supplicant to use the previous PMKID as is for OKC instead of
deriving a new PMKID using an incorrect derivation method when using an
FILS AKM.

Change-Id: I67245eb682a42e44c7955bab8616c800de77bad8
Signed-off-by: Veerendranath Jakkam <vjakkam@codeaurora.org>
CRs-Fixed: 2920431
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: e2e2655ce825a523f3ceca8a066f20533aa04683