Check local supported features for estimating BSS throughputs accurately
Add checks for features supported by the specific hardware mode of the
local device that has the channel for which the throughput is being
estimated instead of assuming the local device supports all optional
features. This is more accurate for cases where the local capabilities
might differ based on the band. In addition, this is in preparation for
extending rate estimates to cover optional VHT and HE features.

CRs-fixed: 2929499
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 1d2118b5095e061b11dd5f08111bde17829f165a
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Change-Id: Ia2efcc9cfa35ed4444fe7e911fa795d2fc4ad050