P2P: Remove 6 GHz channels from full scan if 6 GHz not enabled for P2P
The channels included for the scan to connect to a P2P GO are optimized
such that the P2P GO preferred channel and the common channels are
included for the first few scans followed by a full scan in which all
the channels supported by the local device are included. This results in
P2P client including the 6 GHz channels for the full scan after GO
Negotiation even when 6 GHz channels are not used for the P2P
connection.

Exclude the 6 GHz channels from the full scan if 6 GHz channels are
supported but are not used for P2P connection.

CRs-Fixed: 3083417
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 7dc7b88148de126454439dc8440777cd12ee6bac
Signed-off-by: Sreeramya Soratkal <ssramya@codeaurora.org>
Change-Id: Ie7b77fea012df181874e4a44039c9d4d246bb33a