HE: Disable HE in hostapd_set_freq_params() if driver does not support
Existing logic to disable HE in hostapd_set_freq_params() is to check
he_cap != NULL, but this is not correct as he_cap is defined as a stack
member of hostapd_hw_modes which can't be NULL. Add one more check
!he_cap->he_supported to make sure HE can be disabled if the driver not
support it.

This fixes a case where a driver does not support HE, but hostapd.conf
enables HE/HT40 on the 2.4 GHz band and hostapd failed to start with
error '40 MHz channel width is not supported in 2.4 GHz'.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>

Git-commit: bcaa1ea0848d7f2f48a29294b4427f99ced62832
Git-repo : git://w1.fi/srv/git/hostap.git
CRs-Fixed: 3036800
Change-Id: I4b2acaaa94ba97f57c919f87c26bcb56c3939140