Add support to reconfigure or flush PMKSA cache on interface enable
Update PMKSA cache when interface is disabled and then enabled based on
the new MAC address. If the new MAC address is same as the previous MAC
address, the PMKSA cache entries are valid and hence update the PMKSA
cache entries to the driver. If the new MAC address is not same as the
previous MAC address, the PMKSA cache entries will not be valid anymore
and hence delete the PMKSA cache entries.

Change-Id: Iadb2a75a9958493a890862aa3670ee5fde7ad2b9
Signed-off-by: Veerendranath Jakkam <quic_vjakkam@quicinc.com>
CRs-Fixed: 3057185
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 4775a5f827fa28887db3e12225c56fc2944bac6b