SAE: Do not expire the current PMKSA cache entry
There is no convenient mechanism for reauthenticating and generating a
new PMK during an association with SAE. As such, forced PMK update would
mean having to disassociate and reauthenticate which is not really
desired especially when the default PMKLifetime is only 12 hours.

Postpone PMKSA cache entry expiration of the currently used entry with
SAE until the association is lost. In addition, do not try to force the
EAPOL state machine to perform reauthentication for SAE since that won't
work.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Git-Repo: git://w1.fi/srv/git/hostap.git
Git-Commit: b0f457b6191aa8ee329331ccbe0600326f7b1e2e
CRs-Fixed: 3058348
Change-Id: I76bb78d30f5f82e2682ea1a4cf7a1e27a1051087