P2P: Send response frame on channel where the request is received
The rx_freq of Public Action frame was not maintained by the GO and the
GO always sent the response on the operating channel. This causes
provision discovery failure when a P2P Device is sending a PD Request on
a 2.4 GHz social channel and the GO is responding on a 5 GHz operating
channel.

Save the rx_freq and use it for GO to sent the response. This extends
commit c5cc7a59acb2 ("Report offchannel RX frame frequency to hostapd")
to cover additional frame types.

CRs-Fixed: 3098300
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: f32f99df11ce2bde0039edc003098ee50bb588d4
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Change-Id: Ie996964534b0e19c5026944b18edb7d39f18ad1a