EAPOL: Update PMK length in EAPOL callback to support longer keys
With introduction of newer AKMs, there is a need to update the PMK
length plumbed for the driver based 4-way handshake. To support this,
the current update the PMK length to 48, if the key management type uses
SHA-384. This is needed, e.g., for SUITE-B-192.

Signed-off-by: Vinayak Yadawad <vinayak.yadawad@broadcom.com>
CRs-Fixed: 3294072
Git-repo: git://w1.fi/hostap.git
Git-commit: 909fa448e0131b06b29e3d3bad593b22559df7e6
Change-Id: I84db421815bfbe9e56e634046cfe5c61529381d2