Fix expiration logic for the first PTKSA cache entry
When an entry is added to the PTKSA cache, timer expiration is not set.
Check the list and set the timer expiration when the list is empty also.
When another entry is added to the list, it is placed before the relavant
entry in the order of expiry time of all entries present in the list.

Fixes: a4e36916168a ("WPA: Add PTKSA cache implementation")
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>
Git-commit: 6f8af5974c1b52b473fe910f0bc772e356c6fd98
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 3286704
Change-Id: Idbcaf5a0c5f35e606cdba035b7ffff333b8c749a