PASN: Set keys from PTKSA cache for a valid PTKSA entry
Set secure ranging context to the driver for a valid PTKSA entry in the
PTKSA cache, if available. When the driver requests for PASN keys,
wpa_supplicant configures the keys from its cache if the keys for the
peer derived with the same own MAC address are available in the cache
instead of doing the full PASN handshake again.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>
Git-commit: edd89d6dbb5ae277180aa437460c96c3fabf25f8
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 3286704
Change-Id: I0b9f93e545049e96579f3cf849c23d6ada056902