Validate MAC Address KDE length in the parser
Verify that the MAC Address KDE includes enough data to contain a MAC
address as a part of the parsing function so that each caller would not
need to verify this separately. None of the existing users of this
parser actually needed the MAC address value, so there was not any use
for the length field before. The updated design is more robust for
future uses and gets rid of that unused length field as well.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
CRs-Fixed: 3294072
Git-repo: git://w1.fi/hostap.git
Git-commit: cd392151c59b494663e02e0633e3c6d957aef224
Change-Id: I8e888752c4e9783304d9880569b54ca5b739714f