PASN: Makefile and Android.mk changes for libpasn.so
The wpa_s and hapd context are isolated from the PASN initiator and
responder code and this functionality is now in the separate src/pasn
directory. Add option to build libpasn.so with this functionality. This
library can be used by a Wi-Fi Aware component to support NAN pairing
with other devices.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>
Git-commit: 08abcdf4e7586897ae65d6b24247dc98ecc61ccd
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 3339359
Change-Id: I4959752f58991d9d7ebd686a4a89e3c1a3644bcc