PASN: Store AKMP in the PTKSA cache
PTK is stored in the PTKSA cache following a successful PASN handshake,
however AKMP is removed upon a WPA PASN reset. The PASN handshake is
used in the Wi-Fi Aware R4 specification to define the pairing setup
process. KDK is used to generate a new set of keys, while AKMP is
required for key derivation for pairing. So, keep AKMP in the PTKSA
cache.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>
Git-commit: 4022ffc5db34a2bb43eca7e6245ab0c147c0cf68
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 3339359
Change-Id: Ica4e46ee6e95ccbaca78c170a6ec811ba2d3559d