common: Refactor element defragmentation
Instead of saving the pointers to the fragment elements during parsing
of the frame, append all fragments found right after the element to the
element length. Defragmentation of the element can be done by parsing
appended fragment elements. This approach removes the limit on the
maximum number of fragmented elements supported in a frame.

Signed-off-by: Veerendranath Jakkam <quic_vjakkam@quicinc.com>

CRs-Fixed: 3327346
Git-repo: git://w1.fi/hostap.git
Git-commit: ec03b71ee9995e267437a8a34d551fc1514adb6c
Change-Id: Ie81fbd4e89b15fc78b5465d2777a541fd26e7880