PASN: Fix MIC check not to modify const data
The previous version was using typecasting to ignore const marking for
the input buffer to be able to clear the MIC field for MIC calculation.
That is not really appropriate and could result in issues in the future
if the input data cannot be modified. Fix this by using an allocated
copy of the buffer.

Signed-off-by: Jouni Malinen <j@w1.fi>
Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>
Git-commit: 7e3852407699b7b52516c2af3a75e85caaac1bd1
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-Fixed: 3339359
Change-Id: I9d681f6dd7457a7ae3c2030805532e5a75a7dfa1