mbssid: Configure all BSSes before beacon setup
When multiple BSSID advertisement feature is enabled in IEEE 802.11ax
mode or later, Beacon frames are not transmitted per interface, instead
only one of the interfaces transmits Beacon frames that include one or
more Multiple BSSID elements with configuration for the remaining
interfaces on the same radio.

Change the existing logic such that all configuration details for all
the interfaces are available while building the Beacon frame template
for the transmitting interface itself.

Do not change the flow for the cases where multiple BSSID advertisement
is not enabled.

Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>
CRs-Fixed: 3398041
Git-repo: git://w1.fi/hostap.git
Git-commit: 931e5d4f9e2ea925692ff424e86f8cd6ee9eff8c
Change-Id: I994213f4d615550a56f48cfbb74d68afb926f012